switch-to-configuration: leverage multi-user.target to start units
Instead of manually starting units in switch-to-configuration, we can
leverage the `multi-user.target` unit in order to start all other units
that would normally be started on boot up. This allows for us to remove
all code related to starting units and fixes the problem of starting
units that conflict with one another (since that conflict is resolved by
systemd).