switch-to-configuration-ng: add timeout while waiting for settle
Without a timeout, we (switch-to-configuration) could be stuck in a loop
if we continue receiving D-Bus signals for new systemd events. This
could indicate a problem elsewhere on the machine, however we shouldn't
allow this to cause switch-to-configuration to get stuck.

(cherry picked from commit c776e574d4daca80f68abb3583b6ea27064b3265)