libcap: Fix cross compilation.
Since #442965, the `go run` command that `libcap` does at one point is
no longer able to use the linker included in the go compiler, so it
uses `$CC` instead. But `libcap`'s `Makefile` uses the wrong compiler
for this.

This patch has been submitted to the maintainer upstream. Include it
in nixpkgs for now, until it's accepted and a new release is made.

(cherry picked from commit 315c37461ac0909e8af00fe9d5989a985fb2db15)